Added rest of resistors, including vertical one. Renamed script
